+++ ui/packages/ui/src/app/components/ConnectorTypeComponent.css
