+++ ui/packages/ui/src/app/components/connectorStepHelpers/ConnectorNameTypeHeader.css
