+++ ui/packages/ui/src/app/components/connectorStepHelpers/DataOptionsComponent.tsx
@@ -12,7 +12,7 @@ import { FormikErrors, FormikTouched } from "formik";
 import _ from "lodash";
 import * as React from "react";
 import { formatPropertyDefinitions, PropertyCategory } from "src/app/shared";
-import { FormComponent } from "./formHelpers";
+import { FormComponent } from "../formHelpers";
 
 export interface IDataOptionsComponentProps {
   propertyDefinitions: ConnectorProperty[];
