+++ ui/packages/ui/src/app/components/connectorStepHelpers/FilterTreeComponent.tsx
@@ -9,9 +9,9 @@ import {
 } from "@patternfly/react-core";
 import { CubesIcon, ExclamationCircleIcon } from "@patternfly/react-icons";
 import React from "react";
-import { PageLoader } from ".";
-import { ApiError } from "../shared";
-import { WithLoader } from "../shared/WithLoader";
+import { PageLoader } from "..";
+import { ApiError } from "../../shared";
+import { WithLoader } from "../../shared/WithLoader";
 import "./FilterTreeComponent.css";
 
 export interface IFilterTreeComponentProps {
