+++ ui/packages/ui/src/app/components/connectorStepHelpers/RuntimeOptionsComponent.tsx
@@ -8,7 +8,7 @@ import _ from "lodash";
 import * as React from "react";
 import { formatPropertyDefinitions, PropertyCategory } from "src/app/shared";
 import * as Yup from "yup";
-import { FormComponent } from "./formHelpers";
+import { FormComponent } from "../formHelpers";
 
 export interface IRuntimeOptionsComponentProps {
   propertyDefinitions: ConnectorProperty[];
