+++ ui/packages/ui/src/app/components/index.ts
@@ -1,14 +1,11 @@
 export * from './ApplicationErrorPage';
 export * from './BasicSelectInput';
-export * from './FilterExcludeFieldComponent';
-export * from './FilterInputFieldComponent';
-export * from './FilterTreeComponent';
+export * from './ConnectorIcon';
+export * from './ConnectionPropertiesError';
 export * from './KafkaConnectCluster';
-export * from './NoPreviewFilterField';
 export * from './PageLoader';
 export * from './PageNotFound';
+export * from './connectorStepHelpers';
 export * from './formHelpers';
 export * from './ToastAlertComponent';
-export * from './DataOptionsComponent';
-export * from './PropertiesStepsComponent';
-export * from './RuntimeOptionsComponent';
+export * from './ConnectorTypeComponent';
