+++ ui/packages/ui/src/app/pages/createConnector/CreateConnectorPage.tsx
@@ -1,4 +1,3 @@
-import { Services } from "@debezium/ui-services";
 import {
   Breadcrumb,
   BreadcrumbItem,
