+++ ui/packages/ui/src/app/pages/createConnector/connectorSteps/ConnectorTypeStep.css
