+++ ui/packages/ui/src/app/pages/createConnector/connectorSteps/FilterConfigStep.css
