+++ ui/packages/ui/src/app/pages/createConnector/connectorSteps/index.ts
@@ -1,8 +1,6 @@
-export * from './ConnectorTypeStepComponent';
-export * from './FilterConfigComponent';
+export * from './ConnectorTypeStep';
+export * from './FilterConfigStep';
 export * from './PropertiesStep';
 export * from './DataOptionsStep';
-export * from './ReviewStepComponent';
+export * from './ReviewStep';
 export * from './RuntimeOptionsStep';
-export * from './ConnectorNameTypeHeader';
-export * from './ConnectorTypeComponent';
