+++ ui/packages/ui/src/app/pages/createConnector/federatedModule/DataOptions.tsx
@@ -9,9 +9,9 @@ import {
 import { Formik } from "formik";
 import _ from "lodash";
 import React from "react";
+import { ConnectorNameTypeHeader } from "src/app/components/connectorStepHelpers";
 import { FormComponent } from "../../../components/formHelpers";
 import { formatPropertyDefinitions, PropertyCategory } from "../../../shared";
-import { ConnectorNameTypeHeader } from "../connectorSteps";
 import "./DataOption.css";
 
 export interface IDataOptionsProps {
