+++ ui/packages/ui/src/app/pages/createConnector/federatedModule/FilterConfig.tsx
@@ -14,7 +14,7 @@ import {
   FilterExcludeFieldComponent,
   FilterInputFieldComponent,
   NoPreviewFilterField,
-} from "../../../components";
+} from "../../../components/connectorStepHelpers";
 import {
   ConfirmationButtonStyle,
   ConfirmationDialog,
